requirements.txt
PyMySQL==1.0.2
python-multipart==0.0.5
cryptography==3.4.7
ecdsa==0.17.0
pyasn1==0.4.8
python-jose==3.3.0
rsa==4.7.2
bcrypt==3.2.0 
cffi==1.14.6 
passlib==1.7.4 
pycparser==2.20
asgiref==3.4.1
backcall==0.2.0
click==8.0.1
debugpy==1.4.1
decorator==5.0.9
fastapi==0.68.0
greenlet==1.1.1
h11==0.12.0
parso==0.8.2
pexpect==4.8.0
pickleshare==0.7.5
prompt-toolkit==3.0.19
ptyprocess==0.7.0
pydantic==1.8.2
Pygments==2.9.0
python-dateutil==2.8.2
pytz==2021.1
pyzmq==22.2.1
six==1.16.0
SQLAlchemy==1.4.22
starlette==0.14.2
tornado==6.1
traitlets==5.0.5
typing-extensions==3.10.0.0
uvicorn==0.14.0
wcwidth==0.2.5